06/29/25 2:40.  The design, logo, fonts were changed, elements such as a booking button were added
Unnecessary pages and photos were removed. A new logo was added. Fonts were added: "Lora, Popins, Roboto". The header design was changed, a booking button was added and the login & find a store buttons were removed. A booking button was added to the hero section and js animation was written